+++ docs/conf.py
@@ -42,8 +42,8 @@ extlinks = {
 
 # intersphinx options
 intersphinx_mapping = {'python': ('https://docs.python.org/', None),
-                       'compas': ('http://compas-dev.github.io/main/', None),
-                       'roslibpy': ('http://roslibpy.readthedocs.org/en/latest/', None)}
+                       'compas': ('https://compas.dev/compas/', None),
+                       'roslibpy': ('https://roslibpy.readthedocs.io/en/latest/', None)}
 
 # autodoc options
 autodoc_default_options = {
@@ -59,6 +59,14 @@ autodoc_member_order = 'alphabetical'
 # autosummary options
 autosummary_generate = True
 
+# collect doc versions
+package_docs_root = 'https://gramaziokohler.github.io/compas_fab/'
+
+with open(os.path.join(os.path.dirname(__file__), 'doc_versions.txt'), 'r') as f:
+    version_names = [version.strip() for version in f.readlines()]
+    package_docs_versions = [(version, '{}{}'.format(package_docs_root, version))
+                             for version in version_names if version]
+
 # on_rtd is whether we are on readthedocs.org
 on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
 html_theme = 'compaspkg'
@@ -68,6 +76,8 @@ html_theme_options = {
     "package_title": project,
     "package_version": release,
     "package_repo": 'https://github.com/compas-dev/compas_fab',
+    "package_docs": package_docs_root,
+    "package_old_versions": package_docs_versions
 }
 
 html_split_index = False