+++ docs/examples/03_backends_ros/files/04_attach_mesh_to_ee.py
@@ -6,12 +6,11 @@ import compas_fab
 from compas_fab.backends import RosClient
 from compas_fab.robots import CollisionMesh
 from compas_fab.robots import PlanningScene
-from compas_fab.robots.ur5 import Robot
 
 with RosClient() as client:
-
-    robot = Robot(client)
+    robot = client.load_robot()
     scene = PlanningScene(robot)
+    assert robot.name == 'ur5'
 
     # create collision object
     mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
@@ -21,8 +20,7 @@ with RosClient() as client:
     group = robot.main_group_name
     scene.attach_collision_mesh_to_robot_end_effector(cm, group=group)
 
+    # sleep a bit before removing the tip
     time.sleep(1)
-    scene.remove_attached_collision_mesh('tip')
 
-    # sleep a bit before terminating the client
-    time.sleep(1)
+    scene.remove_attached_collision_mesh('tip')