+++ docs/examples/05_backends_pybullet/files/02_forward_kinematics.py
@@ -1,9 +1,11 @@
-from compas_fab.backends import RosClient
+import compas_fab
+from compas_fab.backends import PyBulletClient
 from compas_fab.robots import Configuration
-from compas_fab.robots.ur5 import Robot
 
-with RosClient() as client:
-    robot = Robot(client)
+with PyBulletClient() as client:
+    urdf_filename = compas_fab.get('universal_robot/ur_description/urdf/ur5.urdf')
+    robot = client.load_robot(urdf_filename)
+
     configuration = Configuration.from_revolute_values([-2.238, -1.153, -2.174, 0.185, 0.667, 0.])
 
     frame_WCF = robot.forward_kinematics(configuration)