+++ CHANGELOG.rst
@@ -13,9 +13,16 @@ Unreleased
 
 **Added**
 
+* **PyBullet integration**: added support for PyBullet client and forward/inverse kinematic solver
 * Added ``ClientInterface``, ``PlannerInterface`` and various backend feature interfaces
 * Added implementations of these interfaces for ROS and V-REP
-* Added client for PyBullet
+* Added ``attributes`` dictionary to ``Robot`` class
+* Added ``compas_fab.robots.Tool.from_t0cf_to_tcf``
+* Added ``compas_fab.robots.Tool.from_tcf_to_t0cf``
+* Added ```joint_names`` as optional parameter for all ``compas_fab.robots.Configuration`` constructors
+* Added ``compas_fab.robots.Configuration.iter_differences``
+* Added ``compas_fab.robots.Configuration.max_difference``
+* Added ``compas_fab.robots.Configuration.close_to``
 
 **Changed**
 
@@ -23,14 +30,26 @@ Unreleased
 * Renamed ``compas_fab.robots.Robot.to_local_coords`` to ``compas_fab.robots.Robot.to_local_coordinates``
 * Renamed ``compas_fab.robots.Robot.to_world_coords`` to ``compas_fab.robots.Robot.to_world_coordinates``
 * Backend clients have been restructured according to the new interfaces
+* Parameter ``backend`` of forward kinematics has been renamed to ``solver``
 * The signatures of all kinematics, motion planning and planning scene management methods have been homogenized across backend clients and within ``Robot``
 * All examples have been updated to reflect these changes
+* Renamed all ``RobotArtist`` implementations to ``RobotModelArtist`` to reflect
+  the fact they depend on ``compas.robots.RobotModel``.
+* Renamed  ``compas_fab.robots.Robot.from_tool0_to_attached_tool`` to ``compas_fab.robots.Robot.from_t0cf_to_tcf``
+* Renamed  ``compas_fab.robots.Robot.from_attached_tool_to_tool0`` to ``compas_fab.robots.Robot.from_tcf_to_t0cf``
+* Changed ROS planning scene methods to be synchronous.
+
 
 **Fixed**
 
 * Attached collision meshes are included in inverse kinematics calculations in ROS
 
 **Deprecated**
+
+* The methods ``forward_kinematics``, ``inverse_kinematics``, ``plan_cartesian_motion`` and ``plan_motion``
+  of ``Robot`` class have been refactored, but a backwards-compatible deprecated version with the old
+  signatures still exists suffixed by ``_deprecated``, e.g. ``forward_kinematics_deprecated``.
+* ``RobotArtist`` are deprecated in favor of ``RobotModelArtist``.
 
 **Removed**